Make all test independent and clean the tests
